more restructuring and refactoring
